Improve debugging for invalid queries through Hvad QuerySet.
When trying to use a non-relation model field as a relation to another model
with a Hvad queryset, this patch improves the error message to help the
developer to fix the error more quickly.

It also improves the error message when you try to assign an object of the
wrong class to a relation, giving the fully-qualified class name, to help
distinguish between many different classes in Python and Django called File.